refactor: API 관련 함수를 별도의 모듈로 분리
